+++ subprojects/mps-gradle-plugin/src/main/kotlin/com/specificlanguages/mps/MpsBuild.kt
@@ -1,15 +1,20 @@
 package com.specificlanguages.mps
 
-import com.specificlanguages.mps.internal.capitalize
 import org.gradle.api.Named
 import org.gradle.api.file.DirectoryProperty
 import org.gradle.api.file.ProjectLayout
 import org.gradle.api.file.RegularFileProperty
 import org.gradle.api.provider.Property
 import org.gradle.api.provider.SetProperty
+import org.gradle.api.tasks.TaskProvider
 import javax.inject.Inject
 
-abstract class MpsBuild : Named {
+abstract class MpsBuild(
+    /**
+     * The task that forks Ant to run the `generate` target of [buildFile].
+     */
+    val generateTask: TaskProvider<out RunAnt>
+) : Named {
     @get:Inject
     protected abstract val layout: ProjectLayout
 
@@ -34,14 +39,9 @@ abstract class MpsBuild : Named {
 
     /**
      * The directory where the build places its artifacts. Typically `$buildDir/artifacts/[build-project-name]`.
+     * `build.layout` Ant property will be set to this directory.
      */
     abstract val buildArtifactsDirectory: DirectoryProperty
-
-    /**
-     * The task that forks Ant to run the `generate` target of [buildFile].
-     */
-    val generateTaskName: String
-        get() = "generate${capitalize(name)}"
 
     /**
      * Other builds that this build depends on. This translates to the dependency on
@@ -54,21 +54,13 @@ abstract class MpsBuild : Named {
     }
 }
 
-abstract class TestBuild : MpsBuild() {
-    /**
-     * The task that forks Ant to run the `check` target of [buildFile]. In MPS-generated Ant files `check` depends on
-     * `assemble` and it is not possible to run `check` alone.
-     */
-    val assembleAndCheckTaskName: String
-        get() = "check${capitalize(name)}"
-}
-
-abstract class MainBuild : MpsBuild() {
+abstract class MainBuild @Inject constructor(
+    generateTask: TaskProvider<out RunAnt>,
     /**
      * The task that forks Ant to run the `assemble` target of [buildFile].
      */
-    val assembleTaskName: String
-        get() = "assemble${capitalize(name)}"
+    val assembleTask: TaskProvider<out RunAnt>
+) : MpsBuild(generateTask) {
 
     /**
      * Indicates whether the build is internal or published (packaged). Published builds have their artifacts included
@@ -81,3 +73,12 @@ abstract class MainBuild : MpsBuild() {
         published.convention(true)
     }
 }
+
+abstract class TestBuild @Inject constructor(
+    generateTask: TaskProvider<out RunAnt>,
+    /**
+     * The task that forks Ant to run the `check` target of [buildFile]. In MPS-generated Ant files `check` depends on
+     * `assemble` and it is not possible to run `check` alone.
+     */
+    val assembleAndCheckTask: TaskProvider<out RunAnt>
+) : MpsBuild(generateTask)